Rename player only context. Add private constructor to LiteFabricFactory
