+++ litecommands-fabric/src/main/java/dev/rollczi/litecommands/fabric/LiteFabricFactory.java
@@ -4,7 +4,7 @@ import dev.rollczi.litecommands.LiteCommandsBuilder;
 import dev.rollczi.litecommands.LiteCommandsFactory;
 import dev.rollczi.litecommands.fabric.argument.PlayerArgument;
 import dev.rollczi.litecommands.fabric.argument.WorldArgument;
-import dev.rollczi.litecommands.fabric.tools.FabricOnlyPlayerContextual;
+import dev.rollczi.litecommands.fabric.context.FabricOnlyPlayerContext;
 import dev.rollczi.litecommands.message.MessageRegistry;
 import dev.rollczi.litecommands.platform.PlatformSettings;
 import net.minecraft.entity.player.PlayerEntity;
@@ -14,7 +14,11 @@ import net.minecraft.server.world.ServerWorld;
 import net.minecraft.text.Text;
 import net.minecraft.world.World;
 
-public class LiteFabricFactory {
+public final class LiteFabricFactory {
+
+    private LiteFabricFactory() {
+    }
+
     @SuppressWarnings("unchecked")
     public static <B extends LiteCommandsBuilder<ServerCommandSource, PlatformSettings, B>> B create() {
         return (B) LiteCommandsFactory.builder(ServerCommandSource.class, new FabricPlatform(new LiteFabricSettings()))
@@ -22,8 +26,8 @@ public class LiteFabricFactory {
                 MessageRegistry<ServerCommandSource> messageRegistry = internal.getMessageRegistry();
 
                 builder
-                    .context(ServerPlayerEntity.class, new FabricOnlyPlayerContextual<>(messageRegistry))
-                    .context(PlayerEntity.class, new FabricOnlyPlayerContextual<>(messageRegistry))
+                    .context(ServerPlayerEntity.class, new FabricOnlyPlayerContext<>(messageRegistry))
+                    .context(PlayerEntity.class, new FabricOnlyPlayerContext<>(messageRegistry))
                     .result(String.class, new StringHandler())
                     .result(Text.class, new TextHandler())
                     .argument(PlayerEntity.class, new PlayerArgument<>(messageRegistry))
@@ -35,4 +39,5 @@ public class LiteFabricFactory {
             })
             ;
     }
+
 }
